requirements/prod.txt
discord.py ~=1.3.3
feedparser ~=5.2.1
pylru ~=1.2.0
